Update festival to the latest version

+++ festival/src/modules/clustergen/HTS_vocoder_me.cc
@@ -183,7 +183,7 @@ void HTS_Vocoder_synthesize_me(HTS_Vocoder_ME * v_me,
                                double *rawdata, HTS_Audio * audio)
 {
   (void) lpf;
-  double x=0;
+  double x=0.0;
   int i, j;
   int k;
   short xs;
@@ -194,7 +194,6 @@ void HTS_Vocoder_synthesize_me(HTS_Vocoder_ME * v_me,
   double xnoise;
   double fxpulse;
   double fxnoise;
-
 
   /* Copy in str's and build pulse and noise shaping filter for this frame */
   for (i = 0; i < v_me->filter_order; i++)
@@ -252,7 +251,7 @@ void HTS_Vocoder_synthesize_me(HTS_Vocoder_ME * v_me,
   }
 
 
-  for (j = 0, i = (IPERIOD + 1) / 2; j < v->fprd; j++)
+    for (j = 0, i = (IPERIOD + 1) / 2; (unsigned int)j < v->fprd; j++)
   {
     if (v->stage == 0) {      /* for MCP */
       if (v->p1 == 0.0)
